fix(release): align updater with current source gates
