+++ developer/sdk/kfd/kfd-1/contract-world.witness.json
@@ -205,9 +205,9 @@
     {
       "name": "kungfu-primitive-catalog",
       "sourcePath": "framework/primitive/kungfu-primitive-catalog.contract.json",
-      "sourceSha256": "4632c930854e495d12d620e531089b5dbf4bdc565bff95960b5f2141ff994f0a",
+      "sourceSha256": "892bf0ff0b7bf8ac32250b340234b6238866dbbca47d30b686459d784a5a8c9e",
       "artifactPath": "config/primitive/kungfu-primitive-catalog.contract.json",
-      "expectedSha256": "4632c930854e495d12d620e531089b5dbf4bdc565bff95960b5f2141ff994f0a",
+      "expectedSha256": "892bf0ff0b7bf8ac32250b340234b6238866dbbca47d30b686459d784a5a8c9e",
       "byteForByte": true
     }
   ],
