+++ docs/evolution/current-authority.md
@@ -11,7 +11,10 @@ contract remains authoritative over this navigation projection.
 | continuation-boundary | Project Cut plus Xinfa Atlas and qualified continuation | [Project Cut and Xinfa continuation](stages/05-project-cut-xinfa.md) | [framework/project-cut/README.md](../../framework/project-cut/README.md)<br>[docs/guides/xinfa-agent-context.md](../guides/xinfa-agent-context.md) |
 | execution-history | native Fact foundation with Episode and Action Geometry | [Native Fact and Action authority](stages/06-native-fact-action.md) | [docs/architecture/fact-episode-action-runtime.md](../architecture/fact-episode-action-runtime.md)<br>[docs/qualification/fact-storage-authority.md](../qualification/fact-storage-authority.md) |
 | extension-composition | native KFX registry with reconciled projections | [Native Work Control and recursive dogfood](stages/08-native-work-control-dogfood.md) | [docs/architecture/kfx-topology.md](../architecture/kfx-topology.md)<br>[docs/profiles/profile-authoring.md](../profiles/profile-authoring.md) |
+| local-product-installation | Shifu selects one same-slot KFD-3 product bundle and native Core owns CLI install, selection, and rollback | [Product Release Cut and updater convergence](stages/10-product-release-cut-updater.md) | [docs/guides/upgrading.md](../guides/upgrading.md)<br>[docs/shifu/schema/local-artifact-catalog-v1.schema.json](../shifu/schema/local-artifact-catalog-v1.schema.json) |
 | primitive-lifecycle | derived Primitive management plane with intake closure | [Native Work Control and recursive dogfood](stages/08-native-work-control-dogfood.md) | [docs/architecture/primitive-management-plane.md](../architecture/primitive-management-plane.md)<br>[docs/architecture/incubation-passport-governance.md](../architecture/incubation-passport-governance.md) |
+| product-release-identity | one immutable Product Release Cut with exact platform slices | [Product Release Cut and updater convergence](stages/10-product-release-cut-updater.md) | [framework/upgrade/kungfu-product-release-cut.contract.json](../../framework/upgrade/kungfu-product-release-cut.contract.json)<br>[docs/adr/KF-ADR-019fabb5-62a0-7b8d-8f8d-6505efdbc239.md](../adr/KF-ADR-019fabb5-62a0-7b8d-8f8d-6505efdbc239.md) |
+| product-update-movement | explicit Cut Transition with lineage, compatibility, migration, rollback, trust, and evidence | [Product Release Cut and updater convergence](stages/10-product-release-cut-updater.md) | [framework/upgrade/kungfu-product-release-cut.contract.json](../../framework/upgrade/kungfu-product-release-cut.contract.json)<br>[framework/upgrade/kungfu-upgrade.contract.json](../../framework/upgrade/kungfu-upgrade.contract.json) |
 | work-authority | one current Work Control identity with an explicit read-only v3 compatibility boundary | [Work Control canonical name](stages/09-work-control-canonical-name.md) | [docs/profiles/work-control.md](../profiles/work-control.md)<br>[config/work-control-naming-boundary.json](../../config/work-control-naming-boundary.json) |
 
 Machine projection: [map.json](map.json).
