+++ scripts/run-runtime-upgrade-tests.mjs
@@ -45,6 +45,14 @@ export function runRuntimeUpgradeTests() {
       'core',
       'tests',
       'python',
+      'test_distribution_update_portability.py',
+    ),
+    path.join(
+      root,
+      'framework',
+      'core',
+      'tests',
+      'python',
       'test_release_channel.py',
     ),
     '-q',
